add: deleteで認証チェックするようにした

+++ app/handler/statuses/delete.go
@@ -3,9 +3,11 @@ package statuses
 import (
 	"encoding/json"
 	"errors"
+	"fmt"
 	"net/http"
 
 	"yatter-backend-go/app/domain/customerror"
+	"yatter-backend-go/app/handler/auth"
 	"yatter-backend-go/app/handler/httperror"
 	"yatter-backend-go/app/handler/request"
 )
@@ -19,7 +21,22 @@ func (h *handler) Delete(w http.ResponseWriter, r *http.Request) {
 		httperror.BadRequest(w, err)
 		return
 	}
+
 	statusRepo := h.app.Dao.Status() // domain/repository の取得
+	status, err := statusRepo.FindWithAccountByID(ctx, id)
+	if err != nil {
+		httperror.InternalServerError(w, err)
+	}
+	if status == nil {
+		httperror.NotFound(w)
+		return
+	}
+	loginAccount := auth.AccountOf(r)
+	if status.Account.ID != loginAccount.ID {
+		httperror.BadRequest(w, fmt.Errorf("Invalid user access"))
+		return
+	}
+
 	if err := statusRepo.DeleteByID(ctx, id); err != nil {
 		if errors.Is(err, customerror.ErrNotFound) {
 			httperror.NotFound(w)
